requirements.txt
google-cloud-storage

pandas